Add index.html, style.css, reset.css, script.js
* index.html
Adds a div container with two paragraphs

* script.js
Adds 64 div elements, gives them classes and appends them to container.

* style.css
Styles div container and the 64 containers created by script.js

Adds media queries, to make the size of the containers change according
to the screen size.

Adds media queries to show show a specific text on screen, according to
the type of the screen. If screen is touchscreen, it shows "Draw something
with your finger". If it's not, it shows "Start drawing with your mouse"

All media queries are commented out, except for one. I did this in order
to be able to work on my code distraction free.

* reset.css
Resets default styles