Modify addClearGridButton(), add addNewGridButton()
* addClearGridButton()
Adds a button with the "Clear Grid" text in it. When button is clicked,
the grid clears and the grid size defaults to 16*16.

* addNewGridButton()
Adds a button with the "New Grid" text in it. When the button is clicked,
asks the user for a new grid and creates the new grid according to the
input the user gave.