Add new functions, modify existing ones, modify README.md
* New functions
- addInputContainer():
Adds a div that has an input field and a button

- addNumberInput():
Creates an input field and its label, and then inserts them both into a
div container. Adds an event listener to input field, so that when a new
grid size is typed in and enter is pressed, we get a new grid, with the
specified size. The colors of the grid remain the same, they don't change.

- addButtonsContainer():
inserts buttonsContainer inside body.

- toggleGridColor():
Toggles between "Change To Black" and "Change To RGB" buttons.
When "Change To Black" button gets clicked, it also gets removed from
parent container, "Change To RGB" button is inserted to parent container
and the grid squares on mouse over change their color to black. When
"Change To RGB" button gets clicked, it also gets removed from parent
container, "Change To Black" button gets inserted to parent container
and the grid squares on mouse over change their color to some random RGB
color.

- generateRGBValue():
Returns a random number between 0 and 255

* Modified existing functions

- addNewGridButton():
Adds a button which on click, changes the size of the grid but keeps the
current color of the grid unchanged. So, if the grid is black, it changes
the size  but keeps the black color. If the grid has random RGB colors,
it changes the grid size, but again keeps the colors.

- addClearGridButton():
It's modified so that it can clear the grid not only when the grid
squares have black color, but when grid squares have random RGB colors
as well.